# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/getUserCount": {
            "get": {
                "tags": [
                    "analysis for public"
                ],
                "summary": "Get Testing",
                "operationId": "testing",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}